feat: add jwt required in update and delete routes, update password, missing key exc and update requirements.txt

+++ app/views/user_lesse_view.py
@@ -1,12 +1,12 @@
-import re
+from app.exc.missing_keys_error import MissingKeys
 from sqlalchemy.exc import IntegrityError
 from app.exc.incorrect_keys_error import IncorrectKeysError
 from flask import Blueprint, request, jsonify
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from http import HTTPStatus
 from app.exc.incorrect_keys_error import IncorrectKeysError
 from app.models.user_lesse_model import UserLesseModel
 from app.services.user_lesse_services import post_user_lesse_by_data, search_user_lesse_by_cpf, delete_user_lesse_by_id, update_user_less_by_id, login_user_lesse
-import ipdb
 
 bp = Blueprint("lesse",__name__, url_prefix="/lesse")
 
@@ -15,10 +15,11 @@ def post_user_lesse_register():
     try:
         data = request.get_json()
         response = post_user_lesse_by_data(data)
-        # ipdb.set_trace()
         return response
 
     except IncorrectKeysError as err:
+        return err.message
+    except MissingKeys as err:
         return err.message
 
     except IntegrityError as err:
@@ -32,21 +33,31 @@ def post_user_lesse_login():
     data = request.get_json()
     try:
         return login_user_lesse(data)
+    except IncorrectKeysError as err:
+        return err.message
+    except MissingKeys as err:
+        return err.message
     except KeyError:
         return {"message": "User not found"}, HTTPStatus.NOT_FOUND
-    ...
+
 
 @bp.patch("/update/<int:user_id>")
+@jwt_required()
 def patch_user_lesse_update(user_id: int):
+    current_user = get_jwt_identity()
     data = request.get_json()
-    try:
-        return update_user_less_by_id(user_id, data)
-    except KeyError as err:
-        return {"message": "You can't change the password, yet!"}, HTTPStatus.UNAUTHORIZED
+    if user_id == current_user['user_id']:
+            return update_user_less_by_id(user_id, data)
+    return {"message": "You need to own the source to modify."}, HTTPStatus.FORBIDDEN
+
 
 @bp.delete("/update/<int:user_id>")
+@jwt_required()
 def del_user_lesse_delete(user_id: int):
-    return delete_user_lesse_by_id(user_id)
+    current_user = get_jwt_identity()
+    if user_id == current_user['user_id']:
+        return delete_user_lesse_by_id(user_id)
+    return {"message": "You need to own the source to modify."}, HTTPStatus.FORBIDDEN
